x/reward/types/message_extract_voucher_test.go
package types

import (
	"testing"

	"titan/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgExtractVoucher_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgExtractVoucher
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgExtractVoucher{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgExtractVoucher{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
